When inserting into the @jobs table, a check is now made to ensure the job is enabled
Made changes to the @jobs insert statement so that it accounts for the HHMMSS format that sysjobhistory's 'duration' column exhibits.
Before it was written as if 'duration' was the total # of seconds.

Next you need to print the backup jobs you're going to be disabling alongside the jobs and there backup files

Tests:

1) Doesn't generate job disabling script if no jobs are eneabled