Added the print statements for the jobs to be disabled Changed the WAITFOR loop so that @job_info is updated after the delay, as it wasn't before.
Changed some of the print statements, appears to be working.

Tentatively, the script is finished.

Started the script that would be run directly after on the secondary instance.
Did some interesting stuff with the table variables that hold the job and db information. May want to look at that before starting on it again.

You ended setting up the print statements when the script is generated explaining what it does and what on