Within the secondary script I changed the @databases table to exclude databases that aren't standby/read-only. This makes it so we can run the script on a server with databases that have already been failed over to.
Within both scripts instead of the transactions being localized to each job that is being disabled, the transaction umbrellas the entire disabling process

Determined if the copy job successfully copied the TLTB (transaction log tail backup) by inserting RESTORE HEADERONLY into a temp table @backupInfo, getting the path for the file from log_shipping_secondary. Checked against the name of the backup to determine if it is the TLTB.

Determined if the restore job successfully restored TLTB by checking the table backupset. Checked against the name of the backup to determine if it is the TLTB.

Still debugging the outputted script, but for the most part it is done.