***Change the d.is_read_only & is_in_standby within the primary logshipping script back to 0 (it was changed to 1 for testing purposes)***
For all the logshipping scripts:

The information for the jobs (schedule, retention periods, etc.) is now going to be sourced from tables (instead of the defaults being taken)
So the backup job for the failover primary will be taken from the original primary's job information tables in msdb.
This means that the scripts for the failover primary and secondary will be ran on the original primary and secondary respectively, in order to have this information available.

The failover primary script has already been transferred over to achieve the above.
The secondary script still needs to get the job info sourced from its databases, however it is already being written in the context of the original secondary so that isn't an issue.