Trying to fix the logins v2 script.
There are 2 parts (that we know of now) that require the version check, which are the 2 insert statements for @principals

The first part was easy, but the second insert has the table variable @temp which is proving difficult to try and pass into dynamic sql.

Trying to use table valued parameters in the 2008+ versions and have the error thrown when run in 2005 as the way to flow the program into the insert code that works with 2005.